pyproject.toml
[tool.poetry]
name = "tg-auth-bot"
version = "0.1.0"
description = ""
authors = ["ivanovsdesign <78684921+ivanovsdesign@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
aiogram = "^3.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
